Create Decorator-Breakfast
They say that breakfast is the most important meal of the day: if that's true, it better be fancy. In this task, we offer you to decorate a slice of bread for your perfect developer's morning. Do not forget about calories!

Sample Input 1:
Sample Output 1:
Bagel, Butter, Butter, Ham, Cheese. kCal: 2890
Bun, Butter, Jam, Jam. kCal: 440