Create Facade-Cinema
The task is to write the correct output using the CinemaFacade.
When we come to the cinema, the first logical step is to get popcorn, so we need to turn on the PopcornPopper. After, we want the facade to turn off the Lights and turn on the Projector. Your task is to write the watchMovie() and finishMovie() methods that should write the output according to our needs.

Sample Input 1:
Sample Output 1:
Get ready to watch a movie...
PopcornPopper on
PopcornPopper popping popcorn!
Lights off
Projector on
We are watching a movie
The End
PopcornPopper off
Lights on
Projector off