Create Strategy-MaxMin
Use the strategy pattern to implement algorithms to find max and min values in a given array. The basic structure of the provided classes is described below: your job is to complete it.
The class Finder represents the general finding algorithm itself. It works according to the specified strategy.
The interface FindingStrategy provides two methods to write new concrete finding strategies: takeOne to get the next value and getDefaultValue to return a value if nothing was found (the given array is empty). Strategies define only the specific details of the finding algorithm.
If the array is empty, the Finder should return Integer.MAX_VALUE in case of finding the min value and Integer.MIN_VALUE in case of finding the max value.